+++ src/complex.rs
@@ -27,7 +27,7 @@ pub fn powi<T>( base: num::complex::Complex<T>, power:usize ) -> num::complex::C
     let temp:T = super::powi( base.re, 2 ) + super::powi( base.im, 2 );
     let mag:T  = super::powi( temp.sqrt(), power );
 
-    let phi:T  = super::atan2( base.im, base.re )*<T>::from_num(power);
+    let phi:T  = super::atan::atan2( base.im, base.re )*<T>::from_num(power);
 
     let real   = mag*super::cos(phi);
     let imag   = mag*super::sin(phi);